pyproject.toml
[tool.poetry]
name = "learning"
version = "0.1.0"
description = ""
authors = ["coder-sabarish <studysabarish@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pillow = "^10.1.0"
requests = "^2.31.0"
pyinstaller = "^6.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
